f/hv/initial_commit

+++ src/Rules/ValidDegreeRule.php
@@ -6,11 +6,17 @@ use HuubVerbeek\ConsistentHashing\Contracts\RuleContract;
 
 class ValidDegreeRule implements RuleContract
 {
-    public function __construct(private readonly float $degree)
+    /**
+     * @param  int  $degree
+     */
+    public function __construct(private readonly int $degree)
     {
         //
     }
 
+    /**
+     * @return bool
+     */
     public function passes(): bool
     {
         return ($this->degree >= 0) && ($this->degree <= 359);
